refactor(DEV-11218): move `getAPIErrorMessage(...)` into `core/utils/` directory
