+++ packages/sdk-react/src/components/tags/TagFormModal/TagFormModal.tsx
@@ -5,7 +5,7 @@ import { toast } from 'react-hot-toast';
 import { useMoniteContext } from '@/core/context/MoniteContext';
 import { MoniteScopedProviders } from '@/core/context/MoniteScopedProviders';
 import { useRootElements } from '@/core/context/RootElementsProvider';
-import { getAPIErrorMessage } from '@/utils/getAPIErrorMessage';
+import { getAPIErrorMessage } from '@/core/utils/getAPIErrorMessage';
 import { yupResolver } from '@hookform/resolvers/yup';
 import type { I18n } from '@lingui/core';
 import { t } from '@lingui/macro';
